+++ filter/bandpass.c
@@ -13,10 +13,10 @@ void filter_bandpass_ticker(void * info) {
     struct filter_bandpass_st *me = (struct filter_bandpass_st *)info;
     
     if ( *(me->freq) != me->lastfreq ) {
-        me->fx = cos(2*PI * *(me->freq) / (float)sample_rate);
-        me->a0 = (1- *(me->res)) * sqrt( *(me->res) * (*(me->res) - 4*me->fx*me->fx + 2) + 1 );
+        me->fx = cosf(2*PI * *(me->freq) / (float)sample_rate);
+        me->a0 = (1- *(me->res)) * sqrtf( *(me->res) * (*(me->res) - 4*me->fx*me->fx + 2) + 1 );
     } else if ( *(me->res) != me->lastres ) {
-        me->a0 = (1- *(me->res)) * sqrt( *(me->res) * (*(me->res) - 4*me->fx*me->fx + 2) + 1 );
+        me->a0 = (1- *(me->res)) * sqrtf( *(me->res) * (*(me->res) - 4*me->fx*me->fx + 2) + 1 );
     }
 
     me->next = me->a0 * *(me->input)
